added schema

+++ src/main/java/com/jobdhundo/Job/Dhundo/repositories/TestimonialRepository.java
@@ -3,5 +3,6 @@ package com.jobdhundo.Job.Dhundo.repositories;
 import com.jobdhundo.Job.Dhundo.entities.Testimonial;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestimonialRepository extends JpaRepository<Testimonial, Long> {
+public interface TestimonialRepository extends JpaRepository<Testimonial, Integer> {
 }
+
